Registrando e excluindo chaves Pix no Banco Central (BCB)

+++ src/main/kotlin/br/com/zup/lucasmiguins/pix/remove/RemoveChavePixService.kt
@@ -1,8 +1,11 @@
 package br.com.zup.lucasmiguins.pix.remove
 
+import br.com.zup.lucasmiguins.integration.bcb.BancoCentralClient
+import br.com.zup.lucasmiguins.integration.bcb.remove.DeletePixKeyRequest
 import br.com.zup.lucasmiguins.pix.ChavePixRepository
 import br.com.zup.lucasmiguins.pix.exceptions.ChavePixClienteNaoEncontradaException
 import br.com.zup.lucasmiguins.shared.validation.ValidUUID
+import io.micronaut.http.HttpStatus
 import io.micronaut.validation.Validated
 import java.util.*
 import javax.inject.Inject
@@ -14,6 +17,7 @@ import javax.validation.constraints.NotBlank
 @Singleton
 class RemoveChavePixService(
     @Inject val repository: ChavePixRepository,
+    @Inject val bcbClient: BancoCentralClient
 ) {
 
     @Transactional
@@ -27,6 +31,12 @@ class RemoveChavePixService(
 
         val chavePix = repository.findByIdAndClienteId(uuidPixId, uuidClienteId)
             .orElseThrow { ChavePixClienteNaoEncontradaException("Não foi encontrada uma chave pix para o cliente") }
+
+        val bcbResponse = bcbClient.deletarChavePix(key = chavePix.chave, DeletePixKeyRequest(chavePix.chave))
+
+        if (bcbResponse.status != HttpStatus.OK) {
+            throw IllegalStateException("Erro ao remover chave Pix no BCB")
+        }
 
         repository.delete(chavePix)
     }
